+++ packages/query-broadcast-client-experimental/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @tanstack/query-broadcast-client-experimental
 
+## 5.90.7
+
+### Patch Changes
+
+- Updated dependencies [[`b4cd121`](https://github.com/TanStack/query/commit/b4cd121a39d07cefaa3a3411136d342cc54ce8fb)]:
+  - @tanstack/query-core@5.90.7
+
+## 5.90.6
+
+### Patch Changes
+
+- Updated dependencies [[`1638c02`](https://github.com/TanStack/query/commit/1638c028df55648995d04431179904371a189772)]:
+  - @tanstack/query-core@5.90.6
+
 ## 5.90.5
 
 ### Patch Changes