+++ packages/solid-query-persist-client/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @tanstack/solid-query-persist-client
 
+## 5.90.10
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-persist-client-core@5.91.6
+  - @tanstack/solid-query@5.90.10
+
+## 5.90.9
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-persist-client-core@5.91.5
+  - @tanstack/solid-query@5.90.8
+
 ## 5.90.8
 
 ### Patch Changes